pyproject.toml
[tool.poetry]
name = "rsi-bot"
version = "0.1.0"
description = ""
authors = ["Kuznetsov Sergey <monteg179@yandex.ru>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["job-queue"], version = "^21.1"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
